pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-comparison-demo"
version = "0.1.0"
description = "Agentic RAG vs GraphRAG side-by-side comparison demo"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Core LLM + orchestration
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    # Vector store
    "qdrant-client>=1.9.0",
    # Async + utils
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.2.0",
    "tenacity>=8.3.0",
    "tiktoken>=0.7.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
graphrag = [
    "neo4j>=5.20.0",
    "spacy>=3.7.0",
    "networkx>=3.3",
    "pyvis>=0.3.2",
]

ui = [
    "gradio>=4.31.0",
    "plotly>=5.22.0",
    "pandas>=2.2.0",
]

observability = [
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "prometheus-client>=0.20.0",
]

api = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "slowapi>=0.1.9",
    "grpcio>=1.64.0",
    "grpcio-tools>=1.64.0",
    "protobuf>=5.27.0",
]

finetuning = [
    "sentence-transformers>=3.0.0",
    "datasets>=2.19.0",
    "transformers>=4.41.0",
    "peft>=0.11.0",
    "accelerate>=0.30.0",
    "ragas>=0.1.13",
]

dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "locust>=2.28.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

all = [
    "rag-comparison-demo[graphrag,ui,observability,api,finetuning,dev]",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=. --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = [
    "core", "pipelines", "orchestrator",
    "graph_db", "api", "ui", "observability",
    "finetuning", "data",
]
